switched to icon buttons being just property additions
