+++ src/demos/list-toolbar/list-toolbar-demo.component.html
@@ -64,30 +64,30 @@
 
     <sky-list-toolbar-view-actions>
       <div class="sky-switch-icon-group">
-        <sky-radio-icon
+        <sky-radio
           icon="table"
           name="iconradiogrouptest"
           value="table"
           [(ngModel)]="iconGroupSelectedValue">
-        </sky-radio-icon>
-        <sky-radio-icon
+        </sky-radio>
+        <sky-radio
           icon="list"
           name="iconradiogrouptest"
           value="list"
           [(ngModel)]="iconGroupSelectedValue">
-        </sky-radio-icon>
-        <sky-radio-icon
+        </sky-radio>
+        <sky-radio
           icon="th-large"
           name="iconradiogrouptest"
           value="cards"
           [(ngModel)]="iconGroupSelectedValue">
-        </sky-radio-icon>
-        <sky-radio-icon
+        </sky-radio>
+        <sky-radio
           icon="map-marker"
           name="iconradiogrouptest"
           value="map"
           [(ngModel)]="iconGroupSelectedValue">
-        </sky-radio-icon>
+        </sky-radio>
       </div>
     </sky-list-toolbar-view-actions>
