+++ src/demos/radio/radio-demo.component.html
@@ -51,42 +51,42 @@
 </h3>
 
 <div>
-    <sky-radio-icon
+    <sky-radio
       icon="info"
       name="iconradiotest"
       radioType="info"
       value="info"
       [(ngModel)]="iconSelectedValue">
-    </sky-radio-icon>
-    <sky-radio-icon
+    </sky-radio>
+    <sky-radio
       icon="strikethrough"
       name="iconradiotest"
       radioType="info"
       value="strikethrough"
       [disabled]="true"
       [(ngModel)]="iconSelectedValue">
-    </sky-radio-icon>
-    <sky-radio-icon
+    </sky-radio>
+    <sky-radio
       icon="star"
       name="iconradiotest"
       radioType="success"
       value="success"
       [(ngModel)]="iconSelectedValue">
-    </sky-radio-icon>
-    <sky-radio-icon
+    </sky-radio>
+    <sky-radio
       icon="exclamation-triangle"
       name="iconradiotest"
       radioType="warning"
       value="warning"
       [(ngModel)]="iconSelectedValue">
-    </sky-radio-icon>
-    <sky-radio-icon
+    </sky-radio>
+    <sky-radio
       icon="ban"
       name="iconradiotest"
       radioType="danger"
       value="danger"
       [(ngModel)]="iconSelectedValue">
-    </sky-radio-icon>
+    </sky-radio>
 </div>
 
 <p>
@@ -98,30 +98,30 @@
 </h3>
 
 <div class="sky-switch-icon-group">
-  <sky-radio-icon
+  <sky-radio
     icon="table"
     name="iconradiogrouptest"
     value="table"
     [(ngModel)]="iconGroupSelectedValue">
-  </sky-radio-icon>
-  <sky-radio-icon
+  </sky-radio>
+  <sky-radio
     icon="list"
     name="iconradiogrouptest"
     value="list"
     [(ngModel)]="iconGroupSelectedValue">
-  </sky-radio-icon>
-  <sky-radio-icon
+  </sky-radio>
+  <sky-radio
     icon="th-large"
     name="iconradiogrouptest"
     value="cards"
     [(ngModel)]="iconGroupSelectedValue">
-  </sky-radio-icon>
-  <sky-radio-icon
+  </sky-radio>
+  <sky-radio
     icon="map-marker"
     name="iconradiogrouptest"
     value="map"
     [(ngModel)]="iconGroupSelectedValue">
-  </sky-radio-icon>
+  </sky-radio>
 </div>
 
 <p>
