+++ src/modules/checkbox/checkbox.component.ts
@@ -66,6 +66,20 @@ export class SkyCheckboxComponent implements ControlValueAccessor {
   @Output()
   public change: EventEmitter<SkyCheckboxChange> = new EventEmitter<SkyCheckboxChange>();
 
+  @Input()
+  public icon: String;
+
+  @Input()
+  public get checkboxType(): string {
+    return this._checkboxType || 'info';
+  }
+  public set checkboxType(value: string) {
+    if (value) {
+      this._checkboxType = value.toLowerCase();
+    }
+  }
+
+  private _checkboxType: string;
   private _checked: boolean = false;
 
   public get inputId(): string {
