+++ src/modules/checkbox/checkbox.module.ts
@@ -6,12 +6,10 @@ import { SkyCheckboxLabelComponent } from './checkbox-label.component';
 import { SkyCheckboxComponent } from './checkbox.component';
 import { SkyResourcesModule } from '../resources';
 import { SkyIconModule } from '../icon';
-import { SkyCheckboxIconComponent } from './checkbox-icon.component';
 
 @NgModule({
   declarations: [
     SkyCheckboxComponent,
-    SkyCheckboxIconComponent,
     SkyCheckboxLabelComponent
   ],
   imports: [
@@ -22,7 +20,6 @@ import { SkyCheckboxIconComponent } from './checkbox-icon.component';
   ],
   exports: [
     SkyCheckboxComponent,
-    SkyCheckboxIconComponent,
     SkyCheckboxLabelComponent
   ]
 })
