+++ src/modules/radio/radio.component.html
@@ -15,7 +15,7 @@
     (blur)="onInputBlur()"
     (ngModelChange)="onRadioChanged($event)"/>
   <span
-    class="sky-switch-control sky-rounded-corners"
+    class="sky-switch-control"
     [ngClass]="{
       'sky-switch-control-icon': icon,
       'sky-rounded-corners': icon,
