+++ src/modules/radio/radio.component.scss
@@ -1,4 +1,4 @@
-.sky-switch-control {
+.sky-switch-control:not(.sky-switch-control-icon) {
   sky-icon {
     font-size: 70%;
   }
