+++ src/scss/_forms.scss
@@ -163,12 +163,12 @@ fieldset {
     border-radius: 0;
 	}
 
-	sky-radio-icon:first-of-type .sky-switch-control, sky-checkbox-icon:first-of-type .sky-switch-control {
+	sky-radio:first-of-type .sky-switch-control-icon, sky-checkbox:first-of-type .sky-switch-control-icon {
 		border-top-left-radius: $sky-border-radius;
 		border-bottom-left-radius: $sky-border-radius;
 	}
 
-	sky-radio-icon:last-of-type .sky-switch-control, sky-checkbox-icon:last-of-type .sky-switch-control {
+	sky-radio:last-of-type .sky-switch-control-icon, sky-checkbox:last-of-type .sky-switch-control-icon {
 		border-top-right-radius: $sky-border-radius;
 		border-bottom-right-radius: $sky-border-radius;
 	}
